WIP: add a nifti 4D reader
This adds the option of importing a 4D nifti file as a multivolume.
The data was tested by converting a diffusion dicom file
using dcm2niix.

The geometry of the imported volume appears mostly correct for the
one volume tested (compared with a nrrd file generated from
the same DICOM data)

But there are many issues that have not been tested
and are not correct.  Among them are:

* the tested file has a QFac of -1, but no other values have been tested

* the time spacing is not used

* the multivolume node attributes are not populated

* the loaded pixel values are not the same even though the image
generally looks similar.  Its possible that rescaling has been
done to one of the images but not the other.